flightSimulator
Imports the data from solarIrradiance, cloudChange and UAVcondition to first run a game where a dot moves along a plane in the specified coordinates. The code in the background calculates the power input & ouput as well as the battery life of the UAV to give the UAV's total distance travelled and the final battery level.